core-app-api: added AppRouter as replacement for app.getRouter()
