+++ .changeset/fuzzy-rivers-search.md
@@ -1,0 +1,5 @@
+---
+'@backstage/core-app-api': minor
+---
+
+Added a new `AppRouter` component that replaces the same component currently created through `app.getRouter()`.
